smartcord
So this is the first draft of smartcord.  It's been used for a few
weeks and my use doesn't seem to bring out any bugs anymore.  There
are still unresolved issues with Tinker's .xyz files: the filename
extension .xyz as well as the extension .001, etc.  Oh well-- it's not
really a descreptive extension, so the only way to fix that is to look
in the files, which I don't want to do right now. Maybe later when I
need to.

darcs-hash:20050501181015-51e1e-fd72e400f9abe966e797bcc1b074b29f47897ebd.gz